finish the workflow before diskChunks

+++ pixels-common/include/physical/Request.h
@@ -5,12 +5,14 @@
 #ifndef PIXELS_REQUEST_H
 #define PIXELS_REQUEST_H
 
+#include <iostream>
+
 class Request {
 public:
-    long queryId;
-    long start;
-    int length;
-    Request(long queryId_, long start_, int length_);
+    uint64_t queryId;
+    uint64_t start;
+    uint64_t length;
+    Request(uint64_t queryId_, uint64_t start_, uint64_t length_);
     int hashCode();
     int comparedTo(Request o);
 };
